pyproject.toml

#  🔨 BUILD SYSTEM
# ==========================================
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


#  📋 PROJECT INFO
# ==========================================
[project]
name = "curator-blogs-system"
version = "0.1.0"
description = "A scallable blog backend system using fastapi"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"

authors = [
    {name = "Abu Bakkar", email = "mdabubakkars182@gmail.com"},
]


dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]


[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8.0",
    "mypy>=1.13",
    "pre-commit>=4.0",
]


[project.urls]
Homepage = "https://github.com/abubakkar-js-dev/curator-frontend"
Repository = "https://github.com/abubakkar-js-dev/curator-frontend"


[project.scripts]
start-server = "app.main:start"


#  🔧 TOOLS CONFIG
# ==========================================

# Ruff — Linter + Formatter
[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

# Pytest
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

# Mypy
[tool.mypy]
python_version = "3.11"
strict = true